+++ R/batchMap.R
@@ -56,7 +56,7 @@ batchMap = function(fun, ..., args = list(), more.args = list(), reg = getDefaul
     as.data.table(x)
   }
 
-  assertRegistry(reg, writeable = TRUE, strict = TRUE)
+  assertRegistry(reg, class = "Registry", writeable = TRUE)
   if (nrow(reg$defs) > 0L)
     stop("Registry must be empty")
   assertFunction(fun)
@@ -95,7 +95,7 @@ batchMap = function(fun, ..., args = list(), more.args = list(), reg = getDefaul
     started     = NA_real_,
     done        = NA_real_,
     error       = NA_character_,
-    memory      = NA_real_,
+    mem.used    = NA_real_,
     resource.id = NA_integer_,
     batch.id    = NA_character_,
     log.file    = NA_character_,