+++ R/syncRegistry.R
@@ -10,8 +10,10 @@
 #' @export
 syncRegistry = function(reg = getDefaultRegistry()) {
   assertRegistry(reg, writeable = TRUE)
-  sync(reg)
-  saveRegistry(reg)
+  if (sync(reg))
+    saveRegistry(reg)
+  else
+    invisible(FALSE)
 }
 
 
@@ -32,7 +34,7 @@ sync = function(reg) {
   updates = rbindlist(updates, fill = TRUE) # -> fill = TRUE for #135
 
   if (nrow(updates) > 0L) {
-    expr = quote(`:=`(started = i.started, done = i.done, error = i.error, memory = i.memory))
+    expr = quote(`:=`(started = i.started, done = i.done, error = i.error, mem.used = i.mem.used))
     reg$status[updates, eval(expr), on = "job.id"]
     if (reg$writeable)
       file.remove.safely(fns[!failed])