+++ R/zzz.R
@@ -50,7 +50,8 @@ batchtools$hooks = list(
     debugme::debugme()
     batchtools$debug = TRUE
   }
-  backports::import(pkgname, c("dir.exists", "hasName", "...length"))
+  backports::import(pkgname, c("dir.exists", "...length", "file.mtime"))
+  backports::import(pkgname, "hasName", force = TRUE)
 } # nocov end
 
 .onUnload = function (libpath) { # nocov start
@@ -59,10 +60,11 @@ batchtools$hooks = list(
 
 .onAttach = function(libname, pkgname) {
   msg = paste(
-    "Breaking change in batchtools v0.9.7:",
+    "Breaking change introduced in batchtools v0.9.6:",
     "The format of the returned data.table of the functions `reduceResultsDataTable()`, getJobTable()`, `getJobPars()`, and `getJobResources()` has changed.",
     "List columns are not unnested automatically anymore.",
-    "To manually unnest tables, batchtools provides the helper function `flatten()` now, e.g. `flatten(getJobPars())`."
+    "To manually unnest tables, batchtools provides the helper function `unwrap()` now, e.g. `unwrap(getJobPars())`.",
+    "The previously introduced helper function `flatten()` will be deprecated due to a name clash with `purrr::flatten()`."
   )
   packageStartupMessage(msg)
 }